add comparison between skorch, pytorch, pytorch_lightning

+++ Crowd_Layer/data_set/dataset.py
@@ -1,6 +1,7 @@
+import numpy as np
 import torch
 from torch.utils.data import Dataset
-import numpy as np
+
 
 def load_dataset():
     data_dir = "./music-multi-annotator-dataset"
@@ -14,6 +15,7 @@ def load_dataset():
     y_test_true = np.load(f'{data_dir}/music-y-true-test.npy')
 
     return X_train, y_train, y_train_true, X_valid, y_valid_true, X_test, y_test_true
+
 
 class MusicDataSet(Dataset):
     def __init__(self, type: str):
@@ -36,14 +38,14 @@ class MusicDataSet(Dataset):
         self.y_train_true = y_train_true
 
     def return_X_y(self):
-        return torch.tensor(self.X), torch.tensor(self.y)
+        return torch.tensor(self.X).float(), torch.tensor(self.y).long()
 
     def return_y_train_true(self):
-        return torch.tensor(self.y_train_true)
+        return torch.tensor(self.y_train_true).float()
 
     def __len__(self):
         return len(self.X)
 
     def __getitem__(self, index):
         X, y = self.X[index], self.y[index]
-        return torch.tensor(X), torch.tensor(y)
+        return torch.tensor(X).float(), torch.tensor(y).long()
